2.1 Color changes
With this commit, the 'c' key will shift the hue (as a measure of
degrees) of the diffuse color of the object by 15 degrees.

This implements two structs:
HCY: A hue-chroma-luma color
RGB: a red-green-blue color
The comments give more information. A conversion function is also
provided to go from HCY to RGB (the reverse is not implemented as it is
never needed).